Fixed bug where entering Preferences would clear saved password
After refactoring Prefs to use dialog fields, set_remember_password was called every time the dialog
closed. prv::password wasn't initialized at that point, meaning the password was always set to "".
This fixes the bug by explicitly clearing it if false is passed instead of setting it every time.

Previous iterations of the Prefs dialog only called the options setters when they were selected, meaning
this never came up as an issue. Toggling the option off would clear the saved password, and toggling it on
would leave it empty until it was re-entered.