HttT S06 Logical Thieves
If you have already killed Agadla by turn 4, instead of offering a choice, the three who appear simply join your forces.

This means it is possible to never get the Thieves, and not be able to recruit them. But you'd have to win before turn 4; so it's not possible unless you're cheating.