Tunnels: fix issues when exit hex is occupied by invisible unit
This can happen when enemy units are hidden under fog or have the hides
ability. Either way, enemy units always blocks the tunnel.

Invisible allied units may or may not block the tunnel, depending on
the setting of pass_allied_units in the [tunnel] tag. It is therefore
necessary to get and check the full teleport map during a unit’s move
through a teleport, but with see_all=true (while the path finding uses
see_all=false).

This commit fixes the behavior and reinstates the ‘failed teleport’
message.