Refactored twesnothd_connection member out of lobby_info
It's only use was to pass refresh_lobby on process_gamelist_diff fail. It makes more sense to
delegate that to the caller - in this case, the lobby. This also means the chatbox widget owns
a twesnothd_connection pointer so the chatbox can be used in situations where no server connection
is present.