Uncomment re-selection guard in tgroup::set_member_active
Without this, multiple selections could occur in the Faction Select dialog. However, somehow the
next option immediately to the disabled one gets selected instead of the first available selection
as expected. While this fixes the FLG dialog, it might have repercussions for the Create Unit dialog.

It's possible this issue has something to do with the ordering of the map. Will have to investigate
further after this.