Add a "simulate lobby activity" plugin
This is the first step towards creating artificial load and testing if the
GUI2 lobby still performs well under load.

I also implemented ability to leave the game in the GUI2 MP staging dialog.
Previously the server continued to believe that the player remains in the
game when he/she closed the MP staging dialog.