MP Staging: refactored handling of settings changes
Instead of immediately regenerating the side configs and pushing them to server in every widget callback (which
is especially expensive in the case of the gold and income sliders), we push updates to the server when the
network callback is called if the state_changed_ flag has been set.

This also cleans up some code from the previous attempt to make both the host and clients utilize this dialog.

Additionally, this fixes gold/income changes not being sent to the server.