Hack: store era addon id in mp game settings
This is so lobby games without the requisite era installed can display a name other than the
era's id, which isn't very nice to look at.

Note that I'm not sure if this is the best or cleanest way to handle this. At some point I'll
need to look more closely at the way addons configs are handled with regards to such games.