No need to redefine PI
Lua, as released, has the correct definition of PI for 'double' (64-bit) floats. It is as precise as possible. If the Boost 128- or 256-bit definitions produce a different value than what Lua provides either (1) we're no longer on 64-bit double floats, (2) there is a typo in Boost, or (3) there is a bug in your compiler.

This reverts the Lua portion of the following commit:

commit f5e673e6443220c57e40cea31d430870b4f73925
Author: fendrin <fabianmueller5@gmx.de>
Date:   Thu Sep 19 01:06:23 2013 +0200

    Replace all kinds of the constant pi representations.

    Replaces every appearance of the circle constant pi I could find with
    the representation from boost/math/constants/constants.hpp.