gui2/tfile_dialog: Add support for a filename extension hint
This allows setting a hint in save mode so that the name portion of the
filename will be selected whenever the user selects a file with a
matching extension on the fileview, and so that whenever the textbox is
cleared by code (e.g. when browsing into directories) it is pre-filled
with the extension with the cursor at the start.

With this, the dialog is now feature-complete compared to the GUI1
version.