fix an issue with text not properly formatted in lua console
gfgtdf noticed that sometimes, if lua generates text or error messages
which looks like pango formatting, it breaks the markup in the lua
console. the problem was that each unit of text generated by lua
needs to be run individually throw font::escape_text. We identified
the problem but for whatever reason didn't fix it back then, and
put a fixme note instead. This is a proper patch that fixes the issue.