desktop: Add functionality to enumerate paths of interest
This allows querying things such as the user's home dir, drive letters
(Windows-only), and game paths such as the data dir, preferences, user
data, and binaries. The results are presented in a format that's
suitable for UI use, with translatable labels used wherever applicable.

At some point there'll be support for listing user-defined bookmarks as
well.