Game Config: added default_color_list= key
This key contains the color range ids of the standard colors in the default order.

Previously, classes such as the connect_engine would have to cast ints to strings and browse for the numerically id-ed
color ranges. Consequently, since there are only 9 ranges identified by that method, 9 became a hardcoded number of
colors in certain places. This key allows for greater control over order and content of the default colors.