Connect Engine: refactored handling of colors
Previously, if any [side] color= were specified at all, a new entry would be added at the beginning of the color
options list, regardless of whether it previously existed as an option or not, and the side_engine::color()
variable was set to 0. This was fine for displaying a color as a value in a comobox, but it looked awful
if one actually looked at the list. Additionally, color() could not be used to apply team coloring (either
via ~RC or ~TC) to anything, since it would always be pointing at 0!

This refactors handling to use more dynamic and sensible methods. A copy of the game_config's default color
list initially populates the color options.

* If any value of color= is already found within that selection, color() will be set to that index value.
* If a numerical value is provided instead, that will be used directly.
* If any non-default color is specified, it will be appended to the end of the color options.

This also removes the getters for pango markup for color options. This is outside the scope of this class and
should be handled by any callers.