MP Staging: fixed crash involving gold/income sliders
For some reason, changing the slider and canceling the dialog caused a crash. Changing the value
while in a networked game also caused a crash. If I remember correctly, this did not occur when the
code used a refence to the side_engine directly, so I'm not sure why this crashed.