fixup! Fix Lua-invoked dialogs not showing before prestart 09d35947f4a05082ec34be292b73a4efc631e1d8
The mentioned commit broke the Lua map generator by removing a parameter from an
overridden method. This commit fixes that by removing it from the superclass as well
and cascading that through.

In addition, override annotations have been added to the generators to guard against
similar mistakes in the future.