MP Create Game: don't update tab contents when regenerating random maps
This moves the tab update code to on_game_select from update_details, since the latter is called when
regenerating random maps and doing so has no possibility of changing tab contents. The only thing that
can do that is selecting a new game. If this ceases to be true in the future, this can be reverted.

This probably should have been done in 80115b11faa3 - in fact, the commit message makes it sound like
it was, but it was not due to an oversight on my part.