MP Options Helper: skip any game, era, or mod with no [options] subtag
Previously, these would still get included in the final options config and result in a [type] id = tag being saved
in prefs with no [option] tags. When loaded in the constructor, this would pollute the options_data_ map, which would
then in turn pollute the resulting config further, ad infinitum.