MP Options Helper: refactor data storage structure
Firstly, this avoids creating duplicate option_source objects in both visible_options and options_data,
and instead makes use of the source id key.

Secondly, it replaces, the sub-map with a config. I don't know if it's any more efficient, but it allows
a simpler transition to option_id = value format after the GUI1 MP screens are dropped.