This removes references to blit_surface and introduces masks for colours
blit_surface is slow compared to the SDL functions and is something we want to remove in order to add hardware rendering.

colours are currently defined by people manually typing out the bit masks, this
lays the groundwork for defining them manually each time. This also starts to acknowledge that alpha is a part of colours.