Slider: comment out slider snap-to-value code (e2d3d585be48)
This seems to cause issue with certain sliders (possibly ones involving negative numbers), and on ones
with large steps the positioner lag is deplorable.

Especially with c70da182b176, having a really smooth slider drag might be preferable. I'm commenting out
the code instead of removing it so it can be revisited.