Fix Lua error when WML assigns a real number to an integer attribute
This regressed in the Lua 5.3 upgrade. LuaL_checkinteger() started to give
an error when the variable isn't an integer. In Lua 5.2 it truncated the
variable instead.

Fortunately we have a macro for modifying attributes, and thus I can fix
all the errors at once.

Maybe a version of the macro that gives an error when the attribute isn't
an integer would be useful as well.