Fix an assertion failure on opening gamestate inspector
The cause was that grids can now call place() for their child widgets
before it has been called for the grid itself. As a result, the size of the
grid is zero, and child widgets which try to fit inside the grid
(see tree_view.cpp:159-161) end up at zero size. In the case of
tree_view_node, attempting to place the node with zero size causes an
assertion failure.

Fixed by updating the size of the grid in request_reduce_* functions.