scons: wesnothd does not need dummy_video.o, breaks wesnoth.org builds
We don't have SDL 2.0.4 on wesnoth.org, so it's not possible to build
dummy_video.o because the config checks fail and SCons doesn't add the
SDL include flags to the environment, and somehow doesn't really warn
that they will be needed for dummy_video.o.

However, since the CMake recipe does not link dummy_video.o into
wesnothd and there doesn't seem to be a reason for wesnothd to require
it (even before we dropped the SDL and SDL_net dependencies from it), we
can probably do away with it.