Reduce dead code
This added a compile-time and a link-time option for gcc and clang. -ffunction-sections causes each function to be placed in a separate section of the object file. -Wl,-gc-functions tells the linker to delete any function sections which are not referenced. The net effect is if a source file has a function which is not used in the final executable program, that function is not present.

While a feature of many modern compilers, function-level linking is a potential source of problems. An example of a potential problem is linking a shared object library (such as a .so .dll or .dylink). In this case, the linker might not see any references to a function and eliminate it from the shared library; in fact, it might eliminate all functions in the shared library.

A better solution, generally used by most major software projects, is to use the "one function, one file" rule. Unfortunately, such a rule seems anethema to wesnoth.