Game Config: fixed red_to_green and blue_to_white returning incorrect values
The issue was that load_config is (for some reason) called twice, and the respective vectors of colors
were simply added on to each time, resulting in incorrect color values when queried.

I figured it would be slightly more efficient to simply assign the vector a new value since the size remains the same.
I'm not sure if a simple vec.clear() call would be more or as efficient, though, not that it really matters at this size.