GUI2/Label: quick fix for label links not being formatted correctly
In 8041a0964e7b the label format was changed to an RGB string, but still passed around verbatim.
This is a quick fix that converts the color into a hex string. The color should probably be stored
as color_t internally, but I'm not sure what to do with the areas that involve formulas operating
on the string.