deprecate needs_select=yes
The behaviour of needs_select=yes event is rather strange and hard to understand:
When a menu items with needs_select=yes it clicked, it first fires the previous select event again but this time in a synced context, so that the select event was fires 2 times total on the playing client and one time at the other clients. It also the breaks the specification in the wiki that select events are not fired in synced contexts.

Also it might happen that the unit stands no longer at the position from where it was selected because it has moved causing unexpected behaviour in this case.

The new way to let other clients know the currents selected unit is the [sync_variable] tag which was added in 1.13.