Stop replacing listboxes with panes with --new-widgets
I don't see the point of such a change. More important, it breaks all
dialogs which search a listbox with find_widget<listbox>() and don't have
a separate code path for --new-widgets, such as the new add-on manager.
In particular, testing the new add-on manager without code changes was
impossible because it was disabled without --new-widgets and failed to
find the add-on list with it.