Drop Down Menu: several improvements to embedded toggle button handling
Any click on a row with a toggle button will now toggle that row's button prior to closing the dialog, unless the
keep_open flag was set, in which case the click will toggle the button but the dialog will not close.

Clicks directly on the toggle buttons themselves are unchanged.