Attempt to fix bugs #25524 and #25525
In order to fix use_map_settings not being set, we create a configure_engine object in the sp game setup
and pass it to the sp options dialog.

I also added checks for the presence of [options] tags in [campaign], as well as check to ensure empty tags
don't cause the dialog to show.

@gfgtdf please confirm fix