Campaign Dialog: fix mod toggling (fixup 66b590b7)
Turns out I made a mistake in that commit. Since the create engine interface relies on toggling a mod, not actually
setting its state directly, my method of toggling a mod if it was selected meant a mod could never been de-selected.

This reverts mod toggling to a callback-based system, using the method implemented in 693e4b4f.