SotA S2: Fix exploit allowing WCs to be created on invalid hexes
The leader should be able to raise corpses only on adjacent hexes.
However, it was possible for the player to right-click on an adjacent
hex then quickly right-click on any other hex, and the corpse would be
raised there instead.

Allowing user interaction to affect a command while it is being
executed may be unintended behavior, and if the engine is changed to
prevent it, this commit can be reverted.