Fix bug #25538: Monte Carlo damage prediction relies on UB
IEEE 754 specifies that dividing a positive number with zero yields
positive infinity, but the C++ language doesn't guarantee IEEE 754
semantics. As a result, UBSan complains about floating-point division by
zero. Wesnoth requires IEEE 754 anyway, but it's still good to get rid of
the UBSan warning.

The code now avoids the division if the divisor is zero.