Small fixup to 42f087abb
The old `tristate_to_bool` function returned the second argument if the value was `t_unset`. boost::tribool can be converted to bool,
but the intermediate state returns false. In this case, the only usecase of `tristate_to_bool` had true as a second argument, and I had
forgotten to explicitly check for the intermediate state when I switched to boost::tribool.

It might be useful to re-add the `tristate_to_bool` function again if conversions to bool become more common, but for now this simple
check works.