Moves drawing of map borders to WML and improves the fake map border
This makes both the standard map border as well as the fake map border terrain be drawn using the same WML rules, which allows them to connect seamlessly with each other.

A dummy IPF ~NO_TOD_SHIFT() is added and used as an easy way to prevent the ToD color-shifting of the border images. This can and probably should be implemented differently in the future, but for now it's a very simple and effective workaround.

This does remove themability of the map border, but that feature is currently unused (even in UMC) and it can be re-implemented differently if the feature is needed. There is also other cleanup still needed regarding themes; the off-map background still remains themable, and the WML for specifying the border tiles still remains in the theme .cfg files.

The call to display::draw_border() which was used to draw the hard-coded map borders is now commented out but not yet removed.