MP Lobby: use exit hook for quit confirmation
By design, this extends the confirmation popup to exiting the lobby with the Log Out button too (previously, only
the Esc key triggered it. However, given the proximity of the button to other control buttons, it's not inconceivable
that someone might click it by accident, so the confirmation is useful.