SotA 16: Fix Archmage having move at start of turn 5
Side X Turn Y occurs before the side's refresh, but the mechanics of refreshing don't occur on turn 1.
This was why the first Archmage had no move but the second one did.

Also adjust positioning so the Archmage is more likely to go to attack the player instead of going for villages or acting confused.