Fix "failed to place a grid" messages when partial relayout is used
The problem was that some widgets which wrap a grid didn't have their
cached sizes thrown away when the grid itself did. As a result, when the
parent grid of such a widget tried partial relayout, it discovered that the
widget fits just fine (according to the cached size) and attempted to lay
it out without shrinking it first. Then the grid found that it was given
much less space than it would prefer, and printed the error message.

Fixed by discarding cached sizes of such widgets.