Refactored handling of the callable variant debug stack
The debug stack is only relevant to callable variants. So instead of continuously passing around a stack reference
to all value classes just so it can be passed back to variant::to_debug_string in the case of recursive calls,
we now keep a static stack member in variant_callable. A second bool flag (clear_stack) has been added to
variant::to_debug_string, to which all recursive value callers should pass false in order to preserve the stack.