Swap the arguments to variant::to_debug_string
The seen stack is an implementation detail that should generally only be
non-null when called recursively from the implementation, so it makes
more sense for it to be the second argument rather than the first.

I did not swap the arguments for variant_value_base::get_debug_string,
because that's entirely internal anyway so it doesn't really matter.