Story Viewer: fixed all floating images drawing in the same spot
Since all the floating images' data was set with a formula that relied on a canvas-wide variable whose value
was constantly changing and no serialization was done, the formulas all evaluated to the same location.

To fix this, we do away with the use of the variables and insert the necessary values directly into the formula
string.