Revert an unnecessary part of 7d2d100
I realized the the theme menu class doesn't contain the *parsed* data - which should indeed be in a config -
it only contains the *ids* of the menu's contents. Therefor, there's no need to store the list of ids in a config,
since it would only have one entry. A string is therefor sufficient.

To enable the input and output vectors being different types, I cleaned up the output initialization loop in
controller_base::show_menu