Cleaned up menu-related code in play_controller::hotkey_handler
For show_menu, there's no need for an explicit check to "save" the autosaves entry anymore, since we add
entries to the item list instead of removing them.

For expand_autosaves, I shouldn't have made the saves list a config just to transform it back to a string.
I only did that since trim_items took a vector of configs. I've now made it templated.