Editor: fix crash when closing map (fixup bf79dbd8)
I had moved the creation of the map_context_refresher prior to deleting a context since I thought it was
necessary to not access invalid memory. Turns out doing that was wrong; doing so caused the crash.

However, this results in a slight change in the code's result: Before, it was taking a ptr to the
current context, modifying the container, and then creating the refresher. This allowed the size_changed_
variable to be correctly initialized. That variable is unused, though, so it makes no difference.