Editor: refactored global labels handling to avoid static pointer to context internals
This should *finally* fix the crash when closing a map (bug #25680). I forgot that the pointer
wouldn't automatically be reset to nullptr when the context to whose labels the pointer was to
was destroyed. Would have been simplest to do that manually, but this makes the code a lot less
confusing and avoid future scenarios where that pointer might be pointing to freed memory.