Editor: fix button overlay disappearing when a GUI2 dialog was open
The problem was that the redraw observer, which handled setting the button states - including the overlays, which
aren't specified in the ThemeWML - was called *after* a draw happened, meaning the buttons were drawn without
overlays properly registered.

This moves the firing of redraw observes before the draw call. The only code that used the redraw ovserver functionality
was the editor, so this shouldn't break anything.