Editor: 'unhide' palette when drawing contents
I'm not sure why, but this allows the enabled state of the to be set properly when switching to a tool
with an empty palette. It seems this function is called after editor_toolkit::hotkey_set_mouse_action
is fired, which already seems to do that, so I'm not sure why it doesn't take effect there, but meh.

Also includes a fixup for when palette_menu_button is fetched. It's a menu button not an action button,
but it probably not havig worked before seems not to have affected anything.