Fix window stack for modeless dialogs
They were not being added to the stack at all, because unlike with
modal dialogs, the window is still visible by the time the show()
function returns. They should have been removed not before reurning
from show() but instead in the hide() function.

This commit only does this for non-tooltip modeless dialogs - dialogs
that allow interaction. The reason for this is that such dialogs could
potentially contain a textbox, which uses the window stack to determine
whether to show a blinking cursor.