WMI Container: removed iterator and container interface
This probably existed to allow simple range-for code in pre-C++11 days when we were using the
BOOST_FOREACH macro. IIRC that required an iterator. However, with C++11 and range-for, it's
much simple to simply specify the `.second` when necessary than having a whole convoluted interface
and iterator extension just to avoid having to type that.

This also changes the return type of 'erase' to bool. It also leaves the empty() and size() functions,
since those might be useful.