Game Events: throw out custom handler_list implementation
This replaces it with the standard std::list interface. I've attempted to make this change as least invasive
as possible, so I have not touched the manager::iteration class. It can probably be refactored further later,
but the crux of this change is simply to allow the removal of the custom smart_list container while keeping
the chance of breakage to a minimum.

The new implementation cannot be an std::forward_list since that only has front-insertion capabilities
while back-insertion is needed.