bi: Add OpenSSL/libcrypto version info
Best 90 minutes I've ever spent in my life. I could've done string
comparisons, but somehow I suspect their format is more likely to change
in the future than the numeric versions.

Also, beware that the upstream documentation (both on their wiki and
their source) does not accurately reflect history -- the version numbers
for releases between 0.9.3 and 0.9.5 were vastly different, and this
version of the algorithm I wrote seems to be the best fit for all of
them.)