gui2/game_version: Try to not overflow tab selector space
The previous commit adding a new library to the Library tab causes a
vertical scrollbar to appear there and somehow make the tab selector
overflow and get a vertical scrollbar. Vultraz suggested this fix and
seems to be aware that it doesn't solve preexisting issues with overlong
tab labels/translations.