GUI2/Styled Widget: use move schematics to greatly reduce the amount of canvas copying
I added some debugging aids, and found that prior to this change, a total of *807* canvas objects
were created by the time the titlescreen was reached, and a new one almost (I think) every time a
tooltip was created.

With this change, the number of new canvas objects created as fallen to *476* by the same point
(a *331* decrease), and no new canvases are created when triggering a tooltip.

I'm not exactly sure how expensive creating a canvas is, though again, it's not a trivial object.
Avoiding hundreds of extra ones should result in some performance boost, at least.