Fixup 9faf17747
It seems while my math in the original commit was correct (both usecases - overflow and amount-to-advance - would
yield in the correct positive amount) I failed to consider the fact that some of the code relied on non-positive
numbers for greater-than comparisons.

To compensate, I've split the differential function in two and made each return 0 instead of negative amounts.
This makes the code clearer while not actually breaking anything.