Fixup ed9c4dff3567
The current resolution values are the effective screen resolution, not the physical ones. So, we want
to exclude resolutions greater than the effective res * the DPI scale factor (ie, the physical resolution),
not the res option * the DPI scale factor. That appears to be because the list of res options is of physical
resolutions so this properly excludes options that wouldn't be able to fit on the screen once scaled.