Allow widget::find_at() to return a scrollbar container
If the pointer wasn't on top of any widget within a scrollbar container,
when GUI2 queried the widget on which a UI event occurred, the answer was
"no widget" and the whole event was discarded.

That was problematic if the player was trying to scroll. It is desired that
scrolling is possible even if the pointer isn't on a widget (as long as
it's somewhere within the scrollbar container).

Now the query will return "scrollbar container" and dispatch the event to
it. That allows the player to scroll anywhere within a scrollbar container.

Also, now the GUI2 event chain includes the target widget itself, not only
its ancestors.

Fixes #1632.