Attempted fix for GUI2 event firing test failures
Commit 315f8496b61db8347b694e2efadfe6ff9ccc7702 broke the test in two ways. First, I intentionally
changed pre and post events to be passed to the target widget as well, not
only its ancestors. The test verified that the events were *not* passed to
the target widget. Fixed by adjusting the test to verify that the target
widget now does receive the event.

Second, I unintentionally stopped the event dispatcher from receiving pre
and post events. Fixed.