wmlunits: Fix issue where the process name becomes "w" instead of "wmlunits"
The basic string type in Python 3 is equivalent to a wchar_t array,
which means that the second byte of every ASCII string is always zero,
causing the string to appear truncated after the first character when
calling C library functions that expect a single or multibyte encoding
(e.g. UTF-8). Just using a byte array does the trick to get a proper
single-byte ASCII string.