GUI2/Window Builder: removed build_widget function and used std::make_shared directly
This function's name was rather misleading. It didn't actually build a widget, just created a an
object (usually a builder, as it was utilized in REGISTER_WIDGET) with the config argument as its
ctor parameters.

Lamdas render the need of a separate function to pass as an std::function argument obsolete.