GUI2/Dispatcher: changed dispatcher signals to take a widget reference as a first parameter
This fixes an issue where functions taking `widget` as a parameter couldn't be bound to a signal function
with the first argument passed via placeholder since `std::bind` requires arguments match exactly when using
the placeholders.

The dispatcher's `this` was cast to `widget` before the signal functions were fired anyway, so this doesn't
really constitute any functional change.