Preprocessor: refactor streambuf management of preprocessor objects
Instead of using this complex system where new processor objects registered themselves as the
current preprocessor object in the base class ctor (this led to freestanding `new` statements
that looked like memory leaks), we now use a deque of unique_ptrs in the streambuf class. This
makes the code a lot clearer since all the memory management is is one place.