Based the FPS counter's displayed value on the longest frametime
The previous behavior caused the counter to often display a very high FPS value due to an occasional long frame causing the delay between following frames to be very short. This commit makes the FPS counter more useful and intuitive for benchmarking purposes by calculating the displayed value based on the longest of the sampled frametimes.