Allow specifying the random shift range for terrain animation times
Previously, terrain graphics [image] random_start= only accepted a boolean value, but this commit makes it optionally also accept an integer value, which will be used as the range of random shift (in milliseconds) that the animation is allowed. This can be used for example to allow animations such as water to play only slightly out of sync.