Revert "Refactor out wesnothd_connection_ptr wrapper class"
This reverts commit 699047766a8852014df20c44d5ed84dadb37015e and fixes #1928. It turns out this made it so
the connection was never closed since it was the wrapper object being destroyed that called the stop()
function.

Since a worker thread needed to hold a shared_ptr to the main ptr in order to ensure it wasn't destroyed
early, we can't just have a shared_ptr of wesnothd_connection directly, since the worker ref means it's
never destroyed.