Fixup 19bbfed again (fixs bug #1941)
TO quote @jyrkive: "prev_def is a pointer to the defendant after the previous fight. The defendant belongs
to bc, it's assigned here. And now when bc is a smart pointer, it frees the battle_context when it goes out
of scope, i.e. when the loop iteration changes. In the next iteration, the stale prev_def is accessed before
a new battle_context is created: use-after-free."