Fix freeze when launching web browser on Linux/macOS
POSIX has a feature to automatically reap zombie processes, which makes it
unnecessary to wait for the child process to exit. Activate the feature
and stop waiting for exit.

"POSIX.1-2001 specifies that if the disposition of SIGCHLD is set to
SIG_IGN or the SA_NOCLDWAIT flag is set for SIGCHLD (see sigaction(2)),
then children that terminate do not become zombies"