GUI2: cleaned up NOTIFY_MODIFIED event triggers
These should be in value setters instead of callback handlers so the events can be fired even when
changes are made programmatically.

Includes a change to the listbox to prevent infinite event recursion when handling toggle panels.
This expands the change from 0dbf686c4ffd to all selectable items.

Also called menu_button::set_selected in menu_button click handler to allow the event firing to be
done in a central place.