Refactor recording of completed difficulties
This stores the data in a completed_at key in the respective [difficulty] subtag instead of as a dedicated key
in the campaign config. This makes the code in Campaign Selection a bit easier to understand and will allow me
to implement the different colored laurels in the Campaign Difficulty dialog more easily.