Fix #1563: graphical glitch when two halos touch
I had added a unrender() call to halo_impl::effect::set_location() in an
attempt to fix haloes from showing in the top-left corner when a Mage of
Light is being moved. It had turned out to be unnecessary in the end,
but I had retained the call because I thought it was the right thing to do.

However, it turned out that removing that call is all it takes to fix bug
#1563.

Thanks to @newfrenchy83 for bringing the fix to my attention.