campaignd: Fix bug causing new uploads to replace server.cfg on disk
This was only temporary and would be later reverted by the periodic
server.cfg commit, but it's still a grave thing and resulted in add-on
packs not actually being written to the correct files, resulting in
missing or outdated contents for uploads performed since commit
ea4f9a4ba2e6d49d67097368fc8b083d3402c629 was pushed to mainline 1.12.

The cause is a trivial copy-paste error that shouldn't have happened.