Fix crash when the AI simulates battles with both swarm and slow
This happened if the AI simulated three fights for an unit that has swarm
and either unit was able to slow in the second fight.

Regression from commit edf750104348841badea50c7d85cd23584d8b1a6.

I didn't take into account that when swarm is used, the probability to stay
unscathed was supposed to be scaled by the probability of the whole combat
slice occurring in the first place. If either unit was able to slow,
complex_fight() calculated the unscaled probability instead, and when they
were added together, the combined probability went way above 100 %. That
triggered an assertion if the AI simulated one more fight for the affected
unit.