DM S09: Fix issues by rebalancing scenario
Give player control of the ghosts and ghouls as they are found.
Update dialog to match ghosts and ghouls joining Delfador.
Enable gold carryover so player doesn't pointlessly wait till last turn.
Lower turn limit, as board is easier with additional units on player side.
Set start gold to 0, so final carryover is lower. With gold and turn limit
changes, it lessens amount in next scenario, but 300 is still achievable,
which is plenty.

Add cave wall to map around tomb. Otherwise a ghost going through the
secret passage causes an akward triggering of the village warning dialog.
Also makes it harder to get to the staff without using the secret passage.