Reconfigure 'select all' for texts in macOS.
Use the idiomatic combination.

Update reference to particular comment line of
'src/widgets/textbox.cpp' in comment with the current line where the
comment is now.

Remove `unsigned` as the type for `copypaste_modifier`, because it was
a little bit less future proof that way, it is better as an
`SDL_Keycode`, which at the moment is `Sint32`, which in turn is
`int32_t` ultimately. Reuse the constant for 'select all' so renaming
it to `modifier_key`.