Remove Unit Box and Widescreen themes
[ci skip]

These were half-baked and mostly badly-designed. The Unit Box theme was never finished, is completely broken,
and if we were to do a major theme refresh we'd probably go in another direction. The Widescreen theme was a
legacy experiment that was never adopted and as of now is basically alsp completely broken.