Revert config unit tests to compare attributes to empty strings again
Reverts a change in commit 2eacb4e1c3c9a14d85aa1602ddc6b248441ed7fa.

The comparison in this unit test compared an *attribute* to an empty
string, not a *string*. Furthermore, the entire point of the test was
to verify that an empty attribute doesn't compare equal to the empty
string. In its old form, the test was just wrong (and failing).