FLG manager: initially select the default faction
Turns out that this is enforced by C++ unit tests (which haven't run
successfully in Travis for a couple of months).

The tests started to fail in commit d864a89a15ee0159b28b5daec492d3070db31032 when random factions
were set to always be the first in the list
(and therefore always initially selected).