Travis CI: globally disable C++ unit tests
The tests haven't been run for a couple of months.
The test executable crashes on startup with the error message

Xlib: extension "RANDR" missing on display ":99.0".

After ten attempts, the test wrapper just gives up and returns success.

So, let's not even pretend that we have C++ unit tests running in Travis.
We just don't. Explicitly disabling the tests makes it more clear.