Revert "Revert "Disallow specifying file paths with backslashes""
This reverts commit 02de08338c0f31b79f1a6604fe7470ee77471323.

Mainline does *not* contain paths with backslashes. If it did, the game
would be broken on GNU/Linux and macOS (and we'd need to fix it anyway).
Instead, it's more likely that we have code somewhere that generates
paths dynamically using backslashes on Windows, and the correct fix is
to change that code to use forward slashes instead.