Revert "Quell divide-by-zero warning"
This reverts commit 07b04775ff3995fca279b15750860f5e79ae2f63.

This code is performance-critical: adding a branch for every pixel in
the surface isn't a good idea. @GregoryLundberg was fine with reverting
the commit, as well. The same is true for the next two reverts.

Fixing Coverity Scan warnings is good and all, but we shouldn't
sacrifice performance for it.