Fixed floating label context dtors sometimes getting stuck in an infinite loop
This fixes #2154. For some reason, after opening help from the File menu and then quitting to desktop,
the number of labels recorded in the label map and context set weren't equal. I'm not exactly sure
why, since the code seems to indicate they should be, and the loop doesn't seem to be hit when only
invoking help with the hotkey (F1).

This just removes the predication of a label being present in the label map for removal from the context
set. I also removed two useless, unused variables that I'm surprised no one's complained about.