Give up on Coverity for this
Coverity wants a cast to double. Add it and it wants a cast to long long. Add THAT and Coverity and GCC both complain.

I suppose there might be a way to rewrite things to make everyone happy. But it's easier to tell Coverity it's a false positive and force the issue closed.