MP: wait at the loading screen until initial lobby data is fetched
This does require some more usability improvements in the future, though. First off, this change was introduced
to get rid of that pause when the lobby is first opened before any contents appears. However, this change only
affects the initial connection, not any subsequent returns to the lobby.

Second, this also means the user doesn't have a way to quit to the titlescreen if the data fetching is taking an
especially long time anymore. It was suggested instead of this, we add some indicator to the lobby that data is
being fetched, but I don't know if that's really feasible with our toolkit. Still, we should add some way to exit
the loading screen if this is taking too long.

Lastly (and building on the above, though this is a more general issue), if at any point the server conks out and
doesn't send the lobby data, or a handshake, or anything, there's no way to exit the loading screen at that point.
Need to come up with a solution for that - either a timeout, or an exit button.